requirements.txt
e2b_code_interpreter
python-dotenv
anthropic
